+++ CHANGELOG.md
@@ -30,6 +30,10 @@ As such, many functions needed to be deprecated and should be replaced according
  - Fixed Output Dock not resizing new rows correctly. - #1260
  - Fixed crash in Process Explorer, when no project is currently open. - #1393
  - Fixed active object changed, while drag and drop an object e.g. into property link properties. - #1397
+ 
+### Added
+ - Horizontal scrolling to the selected object in explorer view.
+   This also fixes jumping of the scrollbar to the left - #1235
 
 ## [2.0.11] - 2025-04-03
 ### Fixed